Fix handling of WEP shared key
For WEP, we need to specify a list of allowed authentication algorithms.
Without those, wpa_supplicant will tell the driver to use an open
authtnetication only. But with this list of authentication algorithms,
mac80211 based drivers will try each and every one of them until one succeeds.
This way we can safely associate with open and shared key WEP networks.

Note that non mac80211 drivers are on their own, but at least they will now be
notified of the authentication algorithms we'd like to try.