Only create new network if it doesn't exists
When calling join_network(), we shouldnt create a new network when
the one we're trying to join is already on the network list.

So this patch looks for a matching network and connect to it, instead
of always creating a 00_00_00_00_00_00 network. Obviously, if the
network is not in the network list we do create the network.