Add checks for networks if attribute is not available
We do actually want to check for NULL pointers, not to prevent g_strcmp0
from crashing, but because we want to compare the 2 pointers when we've
been given one and the network we're looking at has a pointer too.

For example if we call find_network() without an address, it basically
means "find a network with an SSID, and with _any_ address". But then
we're going to compare a NULL pointer with the network's address and
that's gonna be -1. As a consequence, we'll skip this network while
we shouldnt.