Fix broken libiWmxSDK callback parameter handling
libiWmxSDK broke AGAIN the callback model, by passing a 'device_id'
pointer that is not the original the callback was registered
with.

This makes it impossible to obtain the 'wmxsdk' that contains said
'device_id'.

Thus, implemented device_id_to_wmxsdk() by iterating over the list of
known wmxsdks looking for an index match.