Add workaround for DHCP startup failures with WiFi networks
When connecting to a wireless network the supplicant plugin holds
to a network object. But this object is removed from the object table
as a consequence of doing a device disconnect. If a network connect
is in progress when nother connect request comes in (e.g. via D-Bus)
it will irst disconnect which causes the object to be yanked from the
global table. When the first connect request completes the reference
held by the supplicant code is "unregistered" and not usable in starting
up a DHCP client. This leaves it in an unfixable state (restart required).

Workaround is to check in __connman_device_disconnect whether a network
is "connecting" and not disconnect it.