Add workaround for EDNS0 queries that exceed the buffer size
This is a dirty hack which makes it work -- artificially bump the
buffer size up to 4KiB for requests with a smaller buffer size. For
a DNS proxy which serves arbitrary clients this might crash when
it gets a larger response than asked for, but since this is only
serving the local host and knowing that glibc does cope, it is
possible to get away with it for now.