Pass required IEEE 802.1x settings to wpa_supplicant
When trying to connect to a 802.1x service, the attached network is
updated with the 802.1x settings from the config entries, if any. Then
we fetch those settings and send the relevant ones to wpa_supplicant.
Please note that only TLS and PEAP are supported at the moment. Those
two authentication methods account for most of the 802.1x WiFi networks.