Set device powered when bluetooth adapters are powered
When a bluetooth adapters are added, we go through adapter_properties_reply(),
but we dont call connman_device_set_powered() or check_networks() because the
powered variable is false.

We should get the adapter properties upon getting the powered D-Bus reply
message and that will end up calling the right functions.