ofono: Follow registration name changes
As the Name property in org.ofono.NetworkRegistration is sometimes empty
while the network is created, follow the property changes in
org.ofono.NetworkRegistration interface and set network name if the name
changes. Also the network name might change in certain rare cases,
so this needs to be handled in connman side.