Add service statistic data structure
The static values are maintained in the Service object and exposed
through simple accessors.

When a Service object enters the ready state it registers itself at
Counter.

If the Service object is leaving the ready state it will de-register
itself from Counter and consequently it will not be updated anymore.

The user can shorten the update interval when he registers a Counter
object with a shorter interval value.

The statistic is stored in the profile file.
Only the current value is stored, no history.

If there is not Counter object the stats wont be upated. This short
coming will be addressed by the 'data threshold netfilter module'
patches.