Initialize offset counters
The Service object needs the the first update of
the statistic counters to know the offset for its
calculation. Until now, the first timeout of timer
was used to update this values. This fix triggers
an update only for initialiting the offset values
and does not notify the counter clients.