Start/stop statistics depending on LOWER_UP state
The statistic counters should be updated as soon the
interface enters LOWER_UP state ('cable' connected).

Therefore we do not depend on the state of the
Service object, which fixes also the shortcoming
we have currently with private connections. Those
services do not enter the ONLINE state.

At least for ethernet the lower_up callback
has not been called because the service was
created 'too late'.