Save counter values during shutdown
With moving the start/stop function to the lower_up/lower_down
fucntions in service, the statistic was not stored anymore
when connman was shutdown. Reason being that lower_down will
be called only from a RTNL event which is asynchron. So when
the ethernet_disable is called connman doesn't see the
corresponding UP,LOWER_UP event.