Keep start/stop state of statistic timers
When starting a ethernet device the sequence of the
device state (UP, LOWER_UP) from RTNL is different then
by a WiFi device.

For an ethernet device the entering UP and LOWER_UP
arrives together. Whereas for WiFi device we get
a entering UP and leaving LOWER_UP. This results in
a call on lower_down first.

Unfortunatly, the stats_start and stats_stop should
be called in the right order. This fix tracks the
statistics enabled/disabled state.