ofono: use Modem Online property
Modem Online property gets controlled with the cellular device Powered
property.

As the oFono plugin needs to get the Powered state from the device it
gets created when the SIM is ready.

The GPRS networks get added after ConnMan can access the GPRS interface.

As a side effect the oFono modems which do support Online but do not
support GPRS are also visible in ConnMan.

The ConnMan device gets now removed when the SIM is removed or the IMSI
changes.