Try to detect portal even if the location driver is already set
With DHCP renewals the ConnMan state can go from Online to Ready and a
location detect will start. Since the location driver is already set by
then, this will fail and ConnMan's state will be stuck at Ready. Instead
of returning an error, the location finish() hooks is called and a new
portal detection is triggered.