ofono: use new oFono GPRS interfaces
The plugin keeps now a hash table for networks.

The improved oFono GPRS interface (ConnectionManager) relay the complete
network information in ContextAdded signal and in GetContexts reply.  A
connman_network is immediately created for each Internet context.

The kludge used to create a new context in case the initial context list
was empty is removed.

The oFono context interface (ContextManager) has no essential changes,
it is just removed.

Information from the oFono NetworkRegistration interface is now cached
in the connman_device.  Cached registration information is stored in the
newly created connman_network, too. Later changes in the registration
info get propagated to the connman_network.

TODO:
        - handle ConnectionManager.ContextAdded
        - handle ConnectionManager.ContextRemoved