Add verification of the .config file parameters
Connman reads the *.config files in STORAGE_DIR during its boot but
it looks only for parameters it is interrested in.
The configuration parameters syntax is complex and it is very
simple to make a mistake without getting connman warnings. This
leads the user to think that connman understands the configuration
file although it doesn't.

This patch is the code that adds warnings to the logs if connman
reads unknown parameters from the *.config files.