Agent API RequestInput method for requesting passphrases
request_input_passphrase_reply creates and sends an Agent API RequestInput
method call and waits for the method call return. The function expects a
passhphrase_cb_t to be supplied. Only the "Passphrase" dictionary field is
requested by the method call and only the "Passphrase" dictionary field in
the method call return is used.