Remove EDNS0 option
Many routers block EDNS0 DNS requests and that leads to a non working
DNS resolver for ConnMan. This option can now be removed and have dnsproxy
fallling back to DNS over TCP for big DNS replies.
The optimal solution would be to keep this option and disable it on the fly
for nameserver paths rejecting EDNS0 packets. Unfortunately ConnMan is not
able to detect those paths at the moment and so simply falls back to DNS
over TCP.