supplicant: No need for actively remove network upon disconnection
When being disconnected, ConnMan should just wait for wpa_supplicant scan
results instead of trying to be smart about if the disconnected AP is still
around or not.
wpa_supplicant will try to roam and run periodic scan for that purpose. If
the AP is missing after 2 scans, it's removed from the list, and ConnMan
then removes it from the enabled networks.

Fixes BMC#7730 BMC#7734