Merge home and roaming stats file
Instead creating for each service two stats file (home and roaming), just
create one. Each entry in the file is marked with a flag (home or roaming).

The file header contains two additional index which points to the current
home or roaming entry.

Now we have a nice history in the ring buffer.
Additionally this reduces the number of open files by half.