device: Set element driver when probing a device driver
When a device shows up through rtnl before the device driver is created,
the device's element driver pointer will not be set.
The bug is not visible with built-in devices, it only appears with
removable ones or with those that are completely removed from the bus when
being rfkilled.

Fixes BMC#8075