gresolv: Handle POLLERR on DNS UDP socket
If we get an error on the UDP socket, we'll currently go into an endless
loop eating CPU with poll() returning POLLERR and us ignoring it.

This at least hooks things up so that our callback gets called, and we
stop looping. But the callback's handling of POLLERR isn't correct. It
should close the socket and open a new one. As it is, it's just going to
try to use the same broken socket again for the next request. But at least
with this patch it'll stop eating CPU.