gresolv: Don't convert results to strings so early
We're going to want results in sockaddr form for sorting, so let's put
them directly into the array we'll want for the sort process, even though
we haven't actually implemented sorting yet.

We can ditch the support for arbitrary callbacks from the individual
queries, too; there's no need for that. The callback to user code comes
from the *lookup*, having combined the A and AAAA query results.