Fix missing and empty WiFi APs issue after kill/restart
When ConnMan is SIGKILLed and restarted WiFi plugin will reuse the
existing interface and the BSSs list. For BSSs list the next iterator
contains the object path of next BSS but interface_bss_added() wrongly
assumes that the next iterator contains the key/value pairs for the BSS
properties and a GSupplicantNetwork is created with empty fields
resulting in and empty AP list.

Also the iterator is moved to next by both interface_bss_added() and
supplicant_dbus_array_foreach() resulting in not adding alternate
BSSs from the list.

Fix gsupplicant interface_bss_added() to distinguish between BSS's added
with/without the BSS properties.

Fixes BMC#10454 and #11201