ipconfig: Toggle interface IPv6 status properly
This is done so that if service is disconnected, the kernel
created autoconf routes need to be cleared also.
That can be done easily by disabling IPv6 for that interface.
The disabling of IPv6 is only done for autoconfigured interfaces.
When service is connected, the IPv6 is enabled for the used
interface.