ofono: Connect when context activation is requested
ConnMan should connect the corresponding network/service after a
successful response to context activation request. If the context was
already active, connecting has to happen this way instead of waiting for
active property going up (it already is up).

In case the context was inactive before activation, it will not have
network index yet and connecting will happen later when activation is
ready.

This patch removes the reply handler for context inactivation because
it would have no other use than logging.