config: Export a function to load service configs
__connman_config_load_service() function can be used by other parts of
ConnMan core to load service configurations, which can later be used
for service provisioning.

Within config.c, a special field (from_fs) will be used to mark which
configurations originate from the filesystem and which do not. This
information is needed to manage service immutability. Also a special
name "internal" will be used to label the "file" used for non-fs
configurations. A real file will not be created by ConnMan, and it will
be silently ignored if created by someone else. Filesystem storage can
be implemented later if necessary.