session: Introduce append all flag.
When the session is created all current settings should be send to the
Notify Object. One way is to look if the current value
(e.g. EmergencyCall) has changed to the last value. This seems like a
good plan though the code got quite ugly. So let's introduce a flag
for append all. This makes the initializing code quite simple.