dhcp: Invalidate and Release on Network Removal
When removing a network DHCP configuration, invalidate the DHCP
configuration in addition to releasing it.

This addresses a case in which a lower-level network event (e.g. Wi-Fi
disassociation triggered by the driver due to low signal strength)
causes a service disconnect not triggered by connman but then is
recovered BY connman to reuse the prior DHCP lease configuration but
not actually assign it by simply invalidating the old lease
configuration when connman removes the prior network.