loopback: Recover Loopback on Setup when Up but No Address
It is possible to end up in situations in which the loopback interface
is up but has no valid address. In that case, we expect SIOCGIFADDR
will return EADDRNOTAVAIL and valid_loopback should return FALSE such
that connman can add a loopback address back to the interface.

This is a partial fix for bug 21001
<https://bugs.meego.com/show_bug.cgi?id=21001>.