session: Remove deselect_previous_service()
The deselect_and_disconnect() function should be called
from session_changed() and not in the select_and_connect()
code path. All disconnect and connect decision are
initiate from session_changed(). This makes the code flow more
symetric.

Note that we break here a Online -> Offline transision, we are
going to fix this later again.