storage: Add default values to settings file.
On a clean start (without default.profile) create the settings file
with default values. Here is a list of default values:

Offlinemode	= off.
Wifi		= disabled.
bluetooth	= disabled.
ethernet	= disabled.
3G		= disabled.
WiMAX		= disabled.

The reason for offlinemode to be off here is that any changes to
the technology will not be recorded if we start in offlinemode.

Fixes BMC#24285