dnsproxy: Fix can't resolv hostname
When we resolv hostname, we will append the domain name then we may
get question section mismatch error. After we get the DNS reply
we can remove the domain name again. This works well with one host
with only a A record, it may failed when the reply with many cnames
or additional records due to message compression error. But in a
local network we only resolv hostname will not meet these issues.

Modified by Julien Massot according to Samuel's comments.