service: Activate loose mode routing
If more than one service is connected at the same time,
then activate loose mode routing by setting the
/proc/sys/net/ipv4/conf/all/rp_filter to value 2
If the loose mode routing is not activated, then packets
are not routed properly if services are connected to same
subnet.

The original value of rp_filter is restored when the other
services are disconnected and only one service is connected.

For details of rp_filter setting, see Linux kernel file
Documentation/networking/ip-sysctl.txt

Fixes BMC#23606