wifi: Add support for hidden network scan
Implement connman_device_driver scan_hidden function. Supply scan
parameters for the hidden network when calling g_supplicant scan
functionality. When a new scan result is added, compare the SSID
with the one looked for. If detected, connect the hidden network.
When detected or when the scan is done, clear the hidden network
parameters.