technology: Implement Technology Scan and scan start/stop functionality
When Technology Scan D-Bus method call is called, add caller to a scan
list and start a scan for the service type. When a device reports scan
completed, check whether it was the last device scanning and reply to
the callers accordingly.

Also check for scanning status when removing devices and technology
objects.