service: Keep ordering and send all services in ServicesAdded
Keep services ordering and send also already known services
in ServicesAdded signal. Properties are added for new services,
while the dictionary is left empty for already known ones.

When the service is removed, pack the value of "TRUE" to the
removed service hash in order to indicate a removed object path.
The service struct cannot be used, as it is freed immediately
after returning from service_schedule_removed().