ethernet: Do not connect network on LOWER_UP
Ethernet networks should honor the AutoConnect feature as the rest
of the network types.

Just register the network object at the core when we get LOWER_UP.
Then the core can decide when to connect to the network.

There is no behavior change when AutoConnect is enabled, So if the
user plugs the cable the ethernet network will automatically connect
as it was before.

If AutoConnect is disabled, the core will not connect to the network
automatically.