ipconfig: Do not enable IPv6 ipconfigs when loading the properties
When enabling IPv6 when loading the Service properties, the Service
state machine gets out of sync when AutoConnect is disabled.

At this point we should only load the properties and let the
Service state machine decide when to enable the ipconfigs.
And the good news is that the core handles this already.