ipconfig: Disable IPv6 for new interfaces
We don't have yet an ipconfig for this interface, that means
the service.c does not yet know this interface. In order to
be able to get into sync we disable IPv6 auto connection at this point.

If the service state machine decides to connect, it will enable it
again.