test: Remove set-address script
The script is quite useless because it changes the IP address
settings of all found services. The IP settings will be the same
for every service which is probably not what the user wants.
The set-ipv4-method and set-ipv6-method scripts can be used by
the user instead of set-address because those two scripts
will only change settings of one service.