timeserver: Create timeserver list
Create a timeserver list for NTP query. This list consists of
timeserver provided via DHCP and the ones set by user in that
order. We recreate this list everytime the default service changes
or one of the timeserver values change.

Use getaddrinfo to determine if the timeserver is a domain name or
a IP address.